factor out FieldProps type to types file
