+++ protocol-designer/src/components/StepEditForm/fields/DisposalVolumeField.js
@@ -16,7 +16,8 @@ import { selectors as uiLabwareSelectors } from '../../../ui/labware'
 import { getBlowoutLocationOptionsForForm } from '../utils'
 import { TextField } from './TextField'
 
-import type { FieldProps, FieldPropsByName } from './makeSingleEditFieldProps'
+import type { FieldPropsByName } from './makeSingleEditFieldProps'
+import type { FieldProps } from '../types'
 import type { BaseState } from '../../../types'
 
 import styles from '../StepEditForm.css'
