+++ protocol-designer/src/components/StepEditForm/types.js
@@ -1,4 +1,5 @@
 // @flow
+import type { Node } from 'react'
 import type { FormData, StepFieldName } from '../../form-types'
 import type { FieldPropsByName } from './fields/makeSingleEditFieldProps'
 
@@ -17,4 +18,16 @@ export type StepFormProps = {|
 
   focusHandlers: FocusHandlers,
   propsForFields: FieldPropsByName,
+|}
+
+export type FieldProps = {|
+  disabled: boolean,
+  name: string,
+  updateValue: mixed => void,
+  value: mixed,
+  errorToShow: ?string,
+  tooltipContent?: Node,
+  onFieldBlur?: () => mixed,
+  onFieldFocus?: () => mixed,
+  // isIndeterminate?: boolean, // TODO: this will be needed in useBatchEditFieldProps #7222
 |}
